Added settings saving

+++ src/main/kotlin/dev/efnilite/iep/generator/MinSpeedGenerator.kt
@@ -9,11 +9,11 @@ class MinSpeedGenerator : Generator() {
     private var maxSpeed = 0.0
     private var ticksTooSlow = 0
 
-    override val score: Int
+    override val score: Double
         get() {
-            if (startX == 0) return 0
+            if (startX == 0) return 0.0
 
-            return max(0, (players[0].position.x - startX).toInt())
+            return max(0.0, players[0].position.x - startX)
         }
 
     override fun tick() {
