+++ src/main/kotlin/dev/efnilite/iep/generator/SpeedDemonGenerator.kt
@@ -4,8 +4,8 @@ class SpeedDemonGenerator : Generator() {
 
     private var maxSpeedSoFar = 0.0
 
-    override val score: Int
-        get() = maxSpeedSoFar.toInt() // todo add Double
+    override val score: Double
+        get() = maxSpeedSoFar // todo add Double
 
     override fun tick() {
         super.tick()
